+++ test/utils/grid_extension_test.dart
@@ -1,4 +1,4 @@
-import '_prelude.dart';
+import '../_prelude.dart';
 
 void main() {
   group('.isNotEmpty', () {